% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmloo.R
\name{idmloo}
\alias{idmloo}
\title{R function to estimate n illness-death models for leave-one-out data}
\usage{
idmloo(object, data, parallel = FALSE, cpus = NULL, type = NULL, ...)
}
\arguments{
\item{object}{an illness-death model estimate from SmoothHazard package}

\item{data}{data.frame used to estimate object}

\item{parallel}{Logical determinating parallel (=TRUE) or sequential (=FALSE) execution}

\item{cpus}{Numerical amount of CPUs requested for the cluster. If not set, values are total cores of the machines-1.}

\item{type}{Type of cluster. Can be 'SOCK', 'MPI', 'PVM' or 'NWS'. Default is 'SOCK' if parallel=TRUE.}

\item{...}{others parameters}
}
\value{
a list of length N+1 with one indice for leave-one-out of the data and the last
the original estimate of object
}
\description{
R function to estimate n illness-death models for leave-one-out data
}
\examples{
data("Paq1000")
fit.splines <-  idm(formula02=Hist(time=t,event=death)~1,
formula01=Hist(time=list(l,r),event=dementia)~1,
formula12=Hist(time=t,event=death)~1,
method="Splines", kappa=c(200000,200000,20000),
data=Paq1000)
fit.loo.sp=idmloo(object=fit.splines, data=Paq1000, parallel = FALSE, cpus = NULL, type = NULL)


}

