% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo.R
\name{pseudo}
\alias{pseudo}
\title{Pseudo-observations for predictions for an illness-death model}
\usage{
pseudo(object, t, lifeExpect, parallel = FALSE, cpus = NULL, type = NULL,
  ...)
}
\arguments{
\item{object}{an idmloo class object returned by a call to the idmloo function}

\item{t}{time horizon for prediction}

\item{lifeExpect}{logical. if TRUE conmpute life expectancies, i.e., t=Inf}

\item{parallel}{Logical determinating parallel (=TRUE) or sequential (=FALSE) execution}

\item{cpus}{Numerical amount of CPUs requested for the cluster. If not set, values are total cores of the machines-1.}

\item{type}{Type of cluster. Can be 'SOCK', 'MPI', 'PVM' or 'NWS'. Default is 'SOCK' if parallel=TRUE.}

\item{...}{others parameters}

\item{maxtime}{The upper limit of integration for calculations of life expectancies from Weibull parametrizations.}
}
\value{
pseudo
a matrix with each row corresponding to the pseudo-observations of one individual (ordered as in the
the original data set), each column presents the predicted quantities
}
\description{
Computes pseudo-observations for modeling covariates effects on differents quantites of an
illness-death model.
}
\examples{
pseudo.t10=pseudo(fit.loo.w, t=10,lifeExpect=FALSE, parallel = FALSE, cpus=NULL, type=NULL)
}

